requirements.txt
openai
pinecone
streamlit
google.generativeai
python-dotenv